backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.0.0
openstacksdk==1.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
django-filter==23.2
celery==5.2.0
redis==4.5.0
Pillow==9.5.0
python-decouple==3.8
python-dateutil==2.9.0.post0